Shrink main layout to account for when top navbar is used
This prevents the UI stretching past the bottom boundary of the viewport. Originally, the main layout was set to 100% viewport height, but with the top navbar on screen, this effectively pushes the bottom of the UI (mainly the form buttons) down so that a user has to scroll to see them.